Add initial implementation of log-remerge-diff script
This is an attempt at recreating the --remerge-diff flag to log which
was sent as a patch to git, but never accepted[0].

What it does is traverse the log, and when it reaches a merge commit, it
will do the merge again. The diff between the result of that merge, and
the contents in the actual merge commit is shown in the log.

This allows you to see how a merge conflict was resolved, and if any
other changes were included in the merge commit.

[0]: https://www.mail-archive.com/git@vger.kernel.org/msg58059.html